moved table and filter to own file, moved inline styles to component's css files
