+++ ui/src/css/StatusButton.css
@@ -1,0 +1,14 @@
+.status-button{
+  width: 80%;
+  box-sizing: border-box;
+  background-color:red;
+  border: none;
+  display: inline-block;
+  height: 30px;
+  cursor: pointer;
+  border: 1px solid black;
+  border-radius: 10px;
+ 
+}
+
+
